requirements.txt
appdirs==1.4.3
blis==0.2.4
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
ChatterBot==1.0.5
chatterbot-corpus==1.2.0
click==8.0.1
colorama==0.4.3
contextlib2==0.6.0
cymem==2.0.5
distlib==0.3.0
distro==1.4.0
Flask==2.0.1
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
lockfile==0.12.2
MarkupSafe==2.0.1
mathparse==0.1.2
msgpack==0.6.2
murmurhash==1.0.5
mysql-connector-python==8.0.25
nltk==3.6.2
numpy==1.20.3
packaging==20.3
pandas==1.2.5
pep517==0.8.2
Pint==0.17
plac==0.9.6
preshed==2.0.1
progress==1.5
protobuf==3.17.3
pymongo==3.11.4
pyparsing==2.4.6
python-dateutil==2.7.5
pytoml==0.1.21
pytz==2021.1
PyYAML==3.13
regex==2021.4.4
requests==2.22.0
retrying==1.3.3
six==1.14.0
spacy==2.1.9
SQLAlchemy==1.2.19
srsly==1.0.5
thinc==7.0.8
tqdm==4.61.1
urllib3==1.25.8
wasabi==0.8.2
webencodings==0.5.1
Werkzeug==2.0.1
